perbaikan BUK

See merge request ekox/siap2!3

+++ app/Http/Controllers/PenerimaanPajakController.php
@@ -4,149 +4,68 @@ use DB;
 use Session;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use DataTables;
+use App\Libraries\PublicFunction;
 
 class PenerimaanPajakController extends Controller {
 
 	public function index(Request $request)
 	{
-		$aColumns = array('id','nourut','nmunit','nama','nmtrans','nilai','pajak','total');
-		/* Indexed column (used for fast and accurate table cardinality) */
-		$sIndexColumn = "id";
-		/* DB table to use */
-		$sTable = "select  	a.id,
-							lpad(a.nourut,5,'0') as nourut,
-							d.nmunit,
-							e.nama,
-							h.nmtrans,
-							a.nodok as pks,
-							to_char(a.tgdok1,'dd-mm-yyyy') as tgjtempo,
-							a.uraian,
-							nvl(a.nilai_bersih,0) as nilai,
-							b.nmalur||'<br>'||g.nmlevel||'<br>'||c.nmstatus as status,
-							nvl(a.nilai,0)-nvl(a.nilai_bersih,0) as pajak,
-							nvl(a.nilai,0)+nvl(a.ppn,0)+nvl(a.pph21,0)+nvl(a.pph22,0)+nvl(a.pph23,0)+nvl(a.pph25,0) as total
-					from d_trans a
-					left outer join t_alur b on(a.id_alur=b.id)
-					left outer join t_alur_status c on(a.id_alur=c.id_alur and a.status=c.status)
-					left outer join t_unit d on(a.kdunit=d.kdunit)
-					left outer join t_penerima e on(a.id_penerima=e.id)
-					left outer join t_level g on(c.kdlevel=g.kdlevel)
-					left outer join t_trans h on(a.kdtran=h.id)
-					where b.menu=2 and a.thang='".session('tahun')."'
-					order by a.id desc
-					";
-		
-		/*
-		 * Paging
-		 */ 
-		$sLimit = " ";
-		if((isset($_GET['iDisplayStart']))&&(isset($_GET['iDisplayLength']))){
-			$iDisplayStart=$_GET['iDisplayStart']+1;
-			$iDisplayLength=$_GET['iDisplayLength'];
-			$sSearch=$_GET['sSearch'];
-			if ((isset( $iDisplayStart )) &&  ($iDisplayLength != '-1' )) 
-			{
-				$iDisplayEnd=$iDisplayStart+$iDisplayLength-1;
-				$sLimit = " WHERE NO BETWEEN '$iDisplayStart' AND '$iDisplayEnd'";
-			}
-		}
-		
-		/*
-		 * Ordering
-		 */
-		$sOrder = " ";
-		if((isset($_GET['iSortCol_0']))&&(isset($_GET['sSortDir_0']))){
-			$iSortCol_0=$_GET['iSortCol_0'];
-			$iSortDir_0=$_GET['sSortDir_0'];
-			if ( isset($iSortCol_0  ) )
-			{		
-				//modified ordering
-				for($i=0;$i<count($aColumns);$i++){
-					if($iSortCol_0==$i){
-						if($iSortDir_0=='asc'){
-							$sOrder = " ORDER BY ".$aColumns[$i]." DESC ";
-						}
-						else{
-							$sOrder = " ORDER BY ".$aColumns[$i]." ASC ";
-						}
-					}
-				}
-			}
-		}
-		
-		//modified filtering
-		$sWhere="";
-		if(isset($_GET['sSearch'])){
-			$sSearch=$_GET['sSearch'];
-			if((isset($sSearch))&&($sSearch!='')){
-				$sWhere=" where lower(pks) like lower('".$sSearch."%') or lower(pks) like lower('%".$sSearch."%') or
-								lower(nourut) like lower('".$sSearch."%') or lower(nourut) like lower('%".$sSearch."%') or nilai=".$sSearch." ";
-			}
-		}
-		
-		/* Data set length after filtering */
-		$iFilteredTotal = 0;
-		$rows = DB::select("
-			SELECT COUNT(*) as JUMLAH FROM (".$sTable.") qry
-		");
-		$result = (array)$rows[0];
-		if($result){
-			$iFilteredTotal = $result['jumlah'];
-		}
-		
-		/* Total data set length */
-		$iTotal = 0;
-		$rows = DB::select("
-			SELECT COUNT(".$sIndexColumn.") as JUMLAH FROM (".$sTable.") qry
-		");
-		$result = (array)$rows[0];
-		if($result){
-			$iTotal = $result['jumlah'];
-		}
+		$sql = "
+			select  a.id,
+					lpad(a.nourut,5,'0') as nourut,
+					d.nmunit,
+					e.nama,
+					h.nmtrans,
+					a.nodok as pks,
+					to_char(a.tgdok1,'dd-mm-yyyy') as tgjtempo,
+					a.uraian,
+					nvl(a.nilai_bersih,0) as nilai,
+					b.nmalur||'<br>'||g.nmlevel||'<br>'||c.nmstatus as status,
+					nvl(a.nilai,0)-nvl(a.nilai_bersih,0) as pajak,
+					nvl(a.nilai,0)+nvl(a.ppn,0)+nvl(a.pph21,0)+nvl(a.pph22,0)+nvl(a.pph23,0)+nvl(a.pph25,0) as total
+			from d_trans a
+			left outer join t_alur b on(a.id_alur=b.id)
+			left outer join t_alur_status c on(a.id_alur=c.id_alur and a.status=c.status)
+			left outer join t_unit d on(a.kdunit=d.kdunit)
+			left outer join t_penerima e on(a.id_penerima=e.id)
+			left outer join t_level g on(c.kdlevel=g.kdlevel)
+			left outer join t_trans h on(a.kdtran=h.id)
+			where b.menu=2 and a.thang='".session('tahun')."'
+			order by a.id desc
+		";
 
-		/*
-		 * Format Output
-		 */
-		$sEcho="";
-		if(isset($_GET['sEcho'])){
-			$sEcho=$_GET['sEcho'];
-		}
-		$output = array(
-			"sEcho" => intval($sEcho),
-			"iTotalRecords" => $iTotal,
-			"iTotalDisplayRecords" => $iFilteredTotal,
-			"aaData" => array()
-		);
-		
-		$str=str_replace(" , ", " ", implode(", ", $aColumns));
-		
-		$sQuery = "SELECT * FROM ( SELECT ROWNUM AS NO,".$str." FROM ( SELECT * FROM (".$sTable.") ".$sOrder.") ".$sWhere." ) a ".$sLimit." ";
-		
-		$rows = DB::select($sQuery);
-		
-		foreach( $rows as $row )
-		{
-			$aksi='<center>
-						<button type="button" class="btn btn-raised btn-sm btn-icon btn-primary dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false"><i class="fa fa-check"></i></button>
-						<div class="dropdown-menu" x-placement="bottom-start" style="position: absolute; transform: translate3d(0px, 38px, 0px); top: 0px; left: 0px; will-change: transform;">
-							<a id="'.$row->id.'" class="dropdown-item proses" href="javascript:;">Ubah Pajak</a>
-						</div>
-					</center>';
-			
-			$output['aaData'][] = array(
-				$row->no,
-				$row->nourut,
-				$row->nmunit,
-				$row->nama,
-				$row->nmtrans,
-				'<div style="text-align:right;">'.number_format($row->nilai).'</div>',
-				'<div style="text-align:right;">'.number_format($row->pajak).'</div>',
-				'<div style="text-align:right;">'.number_format($row->total).'</div>',
-				$aksi
-			);
-		}
-		
-		return response()->json($output);
+		$query = DB::table(DB::raw("($sql) a"))
+				->selectRaw('a.*');
+
+		$datatables = DataTables::of($query)
+					->addIndexColumn()
+					->editColumn('nilai', function($row){
+						return number_format($row->nilai, 0, ',', '.');
+					})
+					->editColumn('pajak', function($row){
+						return number_format($row->pajak, 0, ',', '.');
+					})
+					->editColumn('total', function($row){
+						return number_format($row->total, 0, ',', '.');
+					})
+					->addColumn('aksi', function($row){
+
+						$aksi='<center>
+								<button type="button" class="btn btn-raised btn-sm btn-icon btn-primary dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false"><i class="fa fa-check"></i></button>
+								<div class="dropdown-menu" x-placement="bottom-start" style="position: absolute; transform: translate3d(0px, 38px, 0px); top: 0px; left: 0px; will-change: transform;">
+									<a id="'.$row->id.'" class="dropdown-item proses" href="javascript:;">Ubah Pajak</a>
+								</div>
+							</center>';
+
+						return $aksi;
+
+					})
+					->rawColumns(['aksi'])
+					->make(true);
+
+		return $datatables;
+
 	}
 	
 	public function pilih(Request $request, $id)
@@ -157,6 +76,7 @@ class PenerimaanPajakController extends Controller {
 					b.nmalur,
 					c.nmunit,
 					d.nama as nmpelanggan,
+					a.kdtran,
 					e.nmtrans,
 					a.nodok as nopks,
 					to_char(a.tgdok,'yyyy-mm-dd') as tgpks,
@@ -350,25 +270,40 @@ class PenerimaanPajakController extends Controller {
 						
 					}
 				}
-				
-				$delete = DB::delete("
-					delete from d_trans_akun
-					where id_trans=?
-				",[
-					$id_trans
-				]);
-					
-				$insert = DB::insert("
-					insert into d_trans_akun(id_trans,kdakun,kddk,nilai,grup)
-					".implode(" union all ", $arr_insert)."
+
+				$rows = DB::select("
+					select	sum(decode(a.kddk,'D',a.nilai,0)) as debet,
+							sum(decode(a.kddk,'K',a.nilai,0)) as kredit
+					from(
+						".implode(" union all ", $arr_insert)."
+					) a
 				");
-				
-				if($insert){
-					DB::commit();
-					return 'success';
+
+				if($rows[0]->debet==$rows[0]->kredit){
+
+					$delete = DB::delete("
+						delete from d_trans_akun
+						where id_trans=?
+					",[
+						$id_trans
+					]);
+						
+					$insert = DB::insert("
+						insert into d_trans_akun(id_trans,kdakun,kddk,nilai,grup)
+						".implode(" union all ", $arr_insert)."
+					");
+					
+					if($insert){
+						DB::commit();
+						return 'success';
+					}
+					else{
+						return 'Simpan pajak gagal!';
+					}
+
 				}
 				else{
-					return 'Simpan pajak gagal!';
+					return 'Jurnal tidak balance, lakukan proses hitung total ulang.';
 				}
 				
 			}
